add common problem result response
