+++ src/WeatherForecast.Host/Features/CurrentWeather/GetCurrentWeatherRequest.cs
@@ -2,9 +2,10 @@
 using MediatR;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using System;
 using System.Threading.Tasks;
 using WeatherForecast.Host.Common;
-using WeatherForecast.Host.Features.ForecastWeather;
+using WeatherForecast.Host.Extensions;
 
 namespace WeatherForecast.Host.Features.CurrentWeather;
 
@@ -16,16 +17,26 @@ public static class GetCurrentWeather
 
         return result.IsSuccess
             ? Results.Ok(result.Value)
-            : Results.StatusCode((int)result.StatusCode);
+            : result.ToProblemResult();
     }
 }
 
 public record GetCurrentWeatherRequest(string City) : IRequest<ApiResponse<GetCurrentWeatherResponse>>;
 
-public class GetCurrentWeatherByCityValidator : AbstractValidator<GetForecastWeatherRequest>
+public class GetCurrentWeatherByCityValidator : AbstractValidator<GetCurrentWeatherRequest>
 {
     public GetCurrentWeatherByCityValidator()
     {
         RuleFor(x => x.City).NotEmpty();
     }
+}
+
+public class GetCurrentWeatherResponse
+{
+    public required string Description { get; init; }
+    public required double Temperature { get; init; }
+    public required double WindSpeed { get; init; }
+    public required DateTime DateTime { get; init; }
+    public required string City { get; init; }
+    public required string Country { get; init; }
 }
