+++ src/WeatherForecast.Host/Features/ForecastWeather/GetForecastWeatherRequest.cs
@@ -3,9 +3,10 @@ using MediatR;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System;
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using WeatherForecast.Host.Common;
-using WeatherForecast.Host.Features.CurrentWeather;
+using WeatherForecast.Host.Extensions;
 
 namespace WeatherForecast.Host.Features.ForecastWeather;
 
@@ -20,16 +21,31 @@ public static class GetForecastWeather
 
         return result.IsSuccess
             ? Results.Ok(result.Value)
-            : Results.StatusCode((int)result.StatusCode);
+            : result.ToProblemResult();
     }
 }
 
 public record GetForecastWeatherRequest(string City, DateTime? DateTime = null) : IRequest<ApiResponse<GetForecastWeatherResponse>>;
 
-public class GetForecastWeatherByCityValidator : AbstractValidator<GetCurrentWeatherRequest>
+public class GetForecastWeatherByCityValidator : AbstractValidator<GetForecastWeatherRequest>
 {
     public GetForecastWeatherByCityValidator()
     {
         RuleFor(x => x.City).NotEmpty();
+    }
+}
+
+public class GetForecastWeatherResponse
+{
+    public required IEnumerable<Forecast> Forecasts { get; init; }
+    public required string City { get; init; }
+    public required string Country { get; init; }
+
+    public class Forecast
+    {
+        public required string Description { get; init; }
+        public required double Temperature { get; init; }
+        public required double WindSpeed { get; init; }
+        public required DateTime DateTime { get; init; }
     }
 }
